requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-Cors==3.0.10
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector==2.2.9
six==1.16.0
Werkzeug==2.3.6
zipp==3.15.0
